requirements.txt
fastapi
loguru
pydantic
requests
ratelimit
python-dotenv
rank-torrent-name
